add servises auth, Ads, user

+++ src/main/java/ru/skypro/homework/dto/AdDTO.java
@@ -10,7 +10,7 @@ import lombok.*;
 public class AdDTO {
     private Long author;
     private String image;
-    private Integer pk;
+    private Long pk;
     private Integer price;
     private String title;
 }
